extract warmUp environment to extensions
