+++ Vostok.Hosting/PublicAPI.Unshipped.txt
@@ -1,3 +1,5 @@
+static Vostok.Hosting.Helpers.WarmUpEnvironmentExtensions.WarmUp(this Vostok.Hosting.Abstractions.IVostokHostingEnvironment environment, Vostok.Logging.Abstractions.ILog log, Vostok.Hosting.VostokHostSettings settings = null) -> void
+Vostok.Hosting.Helpers.WarmUpEnvironmentExtensions
 Vostok.Hosting.Setup.IVostokConsoleLogBuilder.Enable() -> Vostok.Hosting.Setup.IVostokConsoleLogBuilder
 Vostok.Hosting.Setup.IVostokConsoleLogBuilder.IsEnabled.get -> bool
 Vostok.Hosting.Setup.IVostokFileLogBuilder.Enable() -> Vostok.Hosting.Setup.IVostokFileLogBuilder
